+++ src/sfx/sfx.js
@@ -54,7 +54,6 @@ const startTune = document.createElement("audio");
 const raceTune = document.createElement("audio");
 const gameoverFx = document.createElement("audio");
 
-// Make these accessible outside
 export let audioContext;
 export let audioUnlocked = false;
 
@@ -66,7 +65,7 @@ export const unlockAudio = async () => {
     // 1. Try to unlock HTML Audio elements
     try {
         // Quick play attempt on the startTune element
-        startTune.volume = 0.01;
+        startTune.volume = 0.1;
         await startTune.play().catch(e => console.log("First unlock attempt:", e));
         startTune.pause();
         startTune.currentTime = 0;
@@ -89,6 +88,60 @@ export const unlockAudio = async () => {
     
     audioUnlocked = true;
     return true;
+};
+
+// Set up automatic audio unlocking on first interaction
+export const setupAudioUnlock = () => {
+    console.log("Setting up audio unlock for all mobile devices");
+
+    // Create a silent audio element as backup method
+    const silentAudio = document.createElement("audio");
+    silentAudio.setAttribute(
+        "src",
+        "data:audio/mp3;base64,SUQzBAAAAAAAI1RTU0UAAAAPAAADTGF2ZjU4LjIwLjEwMAAAAAAAAAAAAAAA//tUAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWGluZwAAAA8AAAACAAABGwD///////////////////////////////////////////8AAAA8TEFNRTMuMTAwA8MAAAAAAAAAABQgJAUHQQAB9AAAARvMPHBz//////////////////////////////////////////////////////////////////8AAAA"
+    );
+    silentAudio.setAttribute("playsinline", "playsinline");
+    silentAudio.volume = 0.1; // Higher volume to ensure it registers
+
+    // Function to attempt unlocking audio
+    const attemptUnlock = async () => {
+        console.log("Attempting to unlock audio systems from event handler");
+
+        // 1. Try to play the silent sound
+        try {
+            silentAudio
+                .play()
+                .catch((e) => console.log("Silent audio error:", e));
+        } catch (e) {
+            console.log("Silent audio exception:", e);
+        }
+
+        // 2. Call our comprehensive unlockAudio function
+        try {
+            await unlockAudio();
+        } catch (e) {
+            console.log("sfx unlockAudio error:", e);
+        }
+
+        // Remove event listeners after attempting unlock
+        document.removeEventListener("touchstart", attemptUnlock, true);
+        document.removeEventListener("touchend", attemptUnlock, true);
+        document.removeEventListener("click", attemptUnlock, true);
+    };
+
+    // Add listeners in capture phase to handle them first
+    document.addEventListener("touchstart", attemptUnlock, {
+        once: true,
+        capture: true,
+    });
+    document.addEventListener("touchend", attemptUnlock, {
+        once: true,
+        capture: true,
+    });
+    document.addEventListener("click", attemptUnlock, {
+        once: true,
+        capture: true,
+    });
 };
 
 export const initMusicPlayer = (audioTrack, tune, isLooped) => {
@@ -117,6 +170,9 @@ export const initMusicPlayer = (audioTrack, tune, isLooped) => {
 };
 
 export const initialize = () => {
+    // Set up audio unlock automatically
+    setupAudioUnlock();
+    
     return Promise.all([
         initMusicPlayer(startTune, song1, true),
         initMusicPlayer(raceTune, song2, true),
@@ -146,20 +202,30 @@ const FadeOut = (tune, vol = 0) => {
     }
 };
 
+// Check if we're on iOS
+const isIOS = /iPad|iPhone|iPod/.test(navigator.userAgent);
+    
 const FadeIn = (tune, vol = 1) => {
     if (tune._fadeInterval) {
         clearInterval(tune._fadeInterval);
         tune._fadeInterval = null;
     }
 
+
     let playPromise = Promise.resolve();
     if (tune.paused) {
-        tune.volume = 0;
+        // Key change: Start with a small audible volume on iOS
+        tune.volume = isIOS ? 0.1 : 0;
+        
+        // Add playsinline for iOS (just to be safe)
+        tune.setAttribute("playsinline", "playsinline");
         playPromise = tune.play();
     }
 
     playPromise.then(() => {
-        var currentVolume = tune.volume;
+        // Start from current volume
+        var currentVolume = parseFloat(tune.volume);
+        
         if (currentVolume < vol) {
             if (tune._fadeInterval) clearInterval(tune._fadeInterval);
 
@@ -178,6 +244,12 @@ const FadeIn = (tune, vol = 1) => {
         }
     }).catch(e => {
         console.warn("FadeIn play() failed:", e);
+        // Try one more time with higher volume on any failure
+        if (tune.paused) {
+            tune.volume = 1;
+            tune.play().catch(err => console.error("Second play attempt failed:", err));
+        }
+        
         if (tune._fadeInterval) {
             clearInterval(tune._fadeInterval);
             tune._fadeInterval = null;
@@ -201,7 +273,7 @@ const FadeOutIn = (tune1, tune2) => {
             currentVolume = Math.max(0, parseFloat(currentVolume) - 0.1).toFixed(1);
             tune1.volume = currentVolume;
 
-            if (currentVolume <= 0) {
+            if (currentVolume <= 0.1) {
                 tune1.pause();
                 clearInterval(tune1._fadeInterval);
                 tune1._fadeInterval = null;
@@ -258,7 +330,6 @@ export const playTune = async (tune, vol) => {
             break;
         }
         case SFX_START: {
-            startTune.volume = 0;
             startTune.currentTime = 0;
             FadeIn(startTune);
             break;
